Update 6-more-mutations-and-updating-the-store.md
I was receiving an error in the console when I did an upvote: 
Store error: the application attempted to write an object with no provided id but the store already contains an id of Link:cj9sili1zqvn501310yix1dwt for this object. The selectionSet that was trying to be written is:
link {
  votes {
    id
    user {
      id
      __typename
    }
    __typename
  }
  __typename
}

Changing the CREATE_VOTE_MUTATION mutation query fixed the issue.