null object exception issue
In the public class UserRepository , the line 157 user block return a null object,Document doc. This is a way to solve this issue. I know everyone is an excellent Java dev which knows how to solve it in 5 sec. However, there are some new Java devs doesn't know how to solve this issue.