Clarify variable name
Going through this code confused me a little at first, because the purpose of using the 'linkExists' variable is to see if that particular user has already voted on a particular link. So it is not to check if the link itself exists. Additionally, the error thrown specifies that the user has already voted for the link. So I think it would clear things up if that variable was named 'voteExists' instead. I could be misunderstanding something here, but this is how I understood the code.