Update 7-filtering.md
When attempting using the `LinksSearch::Resolver` I kept receiving an error about an undefined method `like` for the ActiveRecordRelation of Links. To fix that, I ended up modifying the scope assignments to be along the lines of `scope.where('description LIKE ?', "%#{value[:description_contains]}%") if value[:description_contains]`. After fixing that and moving forward onto the next parts of this portion of the tutorial, I noticed there was a section about adding the `like` scope to the Link model all the way at the bottom of the page. To prevent this from happening to someone else it would be great to move that block about adding the `like` scope higher up on the page where it's implemented.