Fix formatting of GRAPHENE settings section
The code snippet on the GRAPHENE settings in the section "Configuring django-graphql-jwt" did not show up correctly because the code block was surrounded with normal quotes `'` instead of backticks ` ` `. This is fixed now and should show up correctly.

Also, the module path for the schema was wrong/did not fit the rest of the tutorial: `'SCHEMA': 'mysite.myschema.schema'` instead of `'SCHEMA': 'hackernews.schema.schema'`. This is also fixed.