Fix for #993
As discussed in #993, the `Input` class is not described in this chapter and the question/answer does not correspond with the code examples. The arguments for a mutation in this section are actually defined in the nested `Arguments` class. 
The proposed change reflects this. 

Also, the phrasing of the question was grammatically not ideal and was therefore changed too.