Incorrect answer
Maybe I'm wrong but I think there's been a mistake here with the MCQ. At the end of the chapter, in the "UNLOCK THE NEXT CHAPTER" section, the correct answer is "It is used for authentication". I believe the correct answer should be "It lets resolvers communicate with each other".