Update 6-authentication.md
The 'handler' package was updated such that the "NewDefaultServer" function cannot be found unless the package path is updated to "github.com/99designs/gqlgen/graphql/handler".
Additionally, it was not specified that the server.go file is separated into its own directory. This should be addressed to avoid confusion for those presently following this tutorial.